requirements.txt
Django==3.2.6
djangorestframework==3.12.2
dj-database-url==0.5.0
psycopg2-binary==2.8.6
cx-Oracle==8.2.1
python-dotenv==0.18.0
zeep==4.0.0
celery==5.1.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
git+https://gitlab.com/komtek-python-dev/exchange-framework.git@0.0.1#egg=exchange_framework
gunicorn==20.1.0
redis==3.5.3